Allow reading of ROM space in fixed word length chunks.

+++ src/acorn/A3000.h
@@ -41,9 +41,9 @@ public:
     int InitError();
     int LoadROM(string rom_path);
     int ClockTick();
-    int ReadRAM(uint32_t address, uint8_t *value);
-    int WriteRAM(uint32_t address, uint8_t value);
-    int ReadROM(uint32_t address, uint8_t *value);
+    int ReadRAM(uint32_t address, uint32_t *value);
+    int WriteRAM(uint32_t address, uint32_t value);
+    int ReadROM(uint32_t address, uint32_t *value);
     int Reset();
 
 private:
